Update: 클래스룸 모달창의 데이터 전송(form태그 사용)을 위해 내부 구조 변경 #150
