+++ src/app/classroom/(components)/modal/common/ModalFooter.tsx
@@ -1,11 +1,15 @@
 import LectureSetting from "./LectureSetting";
-import ModalSubmitButton from "./ModalSubmitButton";
 
 const ModalFooter: React.FC = () => {
   return (
     <div className="flex justify-between mb-[-20px]">
       <LectureSetting />
-      <ModalSubmitButton />
+      <button
+        type="submit"
+        className="w-[107px] h-[45px] rounded-[10px] font-bold text-base text-grayscale-20 bg-grayscale-5 hover:text-white hover:bg-primary-80"
+      >
+        업로드
+      </button>
     </div>
   );
 };
