+++ src/app/classroom/(components)/modal/createLecture/AddVideoFileModal.tsx
@@ -4,9 +4,8 @@ import { useDispatch, useSelector } from "react-redux";
 import Image from "next/image";
 import Layout from "../common/Layout";
 import ModalHeader from "../common/ModalHeader";
-import LectureTitle from "../common/LectureTitle";
+import ModalMain from "../common/ModalMain";
 import DropzoneSection from "./DropzoneSection";
-import ModalFooter from "../common/ModalFooter";
 import PageToast from "@/components/PageToast";
 import useClassroomModal from "@/hooks/lecture/useClassroomModal";
 import useVideoFileDrop from "@/hooks/lecture/useVideoFileDrop";
@@ -51,32 +50,33 @@ const AddVideoFileModal: React.FC = () => {
           강의 만들기
         </button>
       </ModalHeader>
-      <LectureTitle />
-      <div className="flex flex-col gap-5 h-72">
-        {videoFile?.type.includes("video") && (
-          <div className="flex gap-3 items-center">
-            <Image
-              src={"/images/fileIcon.svg"}
-              alt={"파일 아이콘"}
-              width={0}
-              height={0}
-              className="w-9 h-auto"
-            />
-            <span className="font-bold text-base text-primary-80">
-              {videoFile?.name}
-            </span>
-            <button
-              type="button"
-              className="ml-auto text-xs text-grayscale-100 hover:font-bold"
-              onClick={handleRemoveVideoFile}
-            >
-              삭제
-            </button>
-          </div>
-        )}
-        <DropzoneSection />
-      </div>
-      <ModalFooter />
+      <ModalMain>
+        <div className="flex flex-col gap-5 h-72">
+          {videoFile?.type.includes("video") && (
+            <div className="flex gap-3 items-center">
+              <Image
+                src={"/images/fileIcon.svg"}
+                alt={"파일 아이콘"}
+                width={0}
+                height={0}
+                className="w-9 h-auto"
+              />
+              <span className="font-bold text-base text-primary-80">
+                {videoFile?.name}
+              </span>
+              <button
+                type="button"
+                className="ml-auto text-xs text-grayscale-100 hover:font-bold"
+                onClick={handleRemoveVideoFile}
+              >
+                삭제
+              </button>
+            </div>
+          )}
+          <DropzoneSection />
+        </div>
+      </ModalMain>
+
       {errorMessage && (
         <PageToast
           toastMsg={errorMessage}
